Added review comment changes
